feat: exact-width integer primitives (byte..ulong)

+++ src/Cvolo.Core/AST/Expressions/IntegerLiteralExpressionSyntax.cs
@@ -3,11 +3,11 @@ using Cvolo.Core.Diagnostics;
 
 namespace Cvolo.Core.AST.Expressions;
 
-public sealed class IntegerLiteralExpressionSyntax(TextSpan span, int value) : ExpressionSyntax(span)
+public sealed class IntegerLiteralExpressionSyntax(TextSpan span, long value) : ExpressionSyntax(span)
 {
 	public override SyntaxKind Kind => SyntaxKind.IntegerLiteralExpression;
 
-	public int Value { get; } = value;
+	public long Value { get; } = value;
 
 	public override IEnumerable<SyntaxNode> GetChildren() => [];
 }
